requirements.txt
jupyter
notebook
ipykernel
openai
python-dotenv
rich
nbformat
pandas
matplotlib
ibm_db
itables
scikit-learn
numpy
spacy
trafilatura
